Add table sorting & move Cluster code down to DataTable
