+++ src/js/tableData/DataTable.js
@@ -2,12 +2,14 @@
 import React, { Component } from 'react';
 import ImageIcon from '../dataset/ImageIcon';
 import { getHeatmapColor } from '../util';
+import clusterfck from 'clusterfck';
 
 export default React.createClass({
 
     getInitialState: function() {
         return {
             showHeatmapColumns: {},
+            sortBy: undefined,
         }
     },
 
@@ -38,6 +40,13 @@ export default React.createClass({
     componentWillUnmount: function() {
         // cleanup plugin
         $(this.refs.dataIcons).selectable( "destroy" );
+    },
+
+    handleSortTable: function(event, name) {
+        event.preventDefault();
+        this.setState({
+            sortBy: name
+        });
     },
 
     handleShowHeatmap: function(event, name) {
@@ -113,8 +122,78 @@ export default React.createClass({
         }
     },
 
+    clusterTableData: function() {
+        // make a list of data for each image from state.tableData
+        // Each item of tableData is already a dict of {imageId: value}
+        let dataKeys = Object.keys(this.props.tableData);
+        if (dataKeys.length < 2) return;
+        // make dict of {imageId: [v1, v2, v3]}.
+        let imageIds = this.props.imgJson.map(i => i.id);
+        let toCluster = imageIds.map(iid => {
+            return dataKeys.map(key => this.props.tableData[key][iid]);
+        });
+        // Make a lookup of tableDataKey :image_id so that when we get clustered tableDAta
+        // we can work out which image each 'row' of data came from (since we can't
+        // include image ID in the cluster data)
+        // We know the 'toCluster' list is in same order as imageIDs just now...
+        let clusterLookup = imageIds.reduce((prev, iid, index) => {
+            let dataString = toCluster[index].join(",");
+            prev[dataString] = iid;
+            return prev;
+        }, {});
+        // let toCluster = [[255, 255, 240],
+        //     [20, 120, 102],
+        //     [250, 255, 253],
+        //     [100, 54, 300]];
+        let threshold = 25000;
+        var clusters = clusterfck.hcluster(toCluster, clusterfck.EUCLIDEAN_DISTANCE,
+            clusterfck.AVERAGE_LINKAGE, threshold);
+        let orderedResults = this.traverseCluster(clusters);
+
+        let orderedImageIds = orderedResults.map(res => clusterLookup[res.join(",")]);
+        return orderedImageIds;
+    },
+
+    traverseCluster: function(clusters) {
+
+        let out = [];
+        function traverseNode(node) {
+            if (node.left && node.right) {
+                traverseNode(node.left);
+                traverseNode(node.right);
+            } else {
+                out.push(node.value);
+            }
+        }
+        clusters.forEach(n => {traverseNode(n)});
+        return out;
+    },
+
     render() {
         let {imgJson, iconSize, tableData} = this.props;
+
+        if (this.state.sortBy != undefined) {
+            let orderedImageIds;
+            if (this.state.sortBy === "cluster") {
+                orderedImageIds = this.clusterTableData();
+            } else {
+                let colDataToSort = this.props.tableData[this.state.sortBy];
+                // convert {iid: value} to [value, iid]
+                let dataList = [];
+                for (var iid in colDataToSort) {dataList.push([iid, colDataToSort[iid]])};
+                dataList.sort((a, b) => a[1] < b[1] ? -1 : 1);
+                orderedImageIds = dataList.map(l => l[0]);
+            }
+            if (orderedImageIds) {
+                // sort filteredImages by ID...
+                // create lookup {iid: image}
+                let imgLookup = imgJson.reduce((prev, img) => {
+                    prev[img.id] = img;
+                    return prev;
+                }, {});
+                imgJson = orderedImageIds.map(iid => imgLookup[iid]);
+            }
+        }
 
         let columnNames = Object.keys(tableData);
 
@@ -135,11 +214,17 @@ export default React.createClass({
                 <table className="parade_dataTable">
                     <tbody>
                     <tr>
-                        <td>Thumb</td>
+                        <td>
+                            <a onClick={(event) => {this.handleSortTable(event, "cluster")}}>
+                                Cluster
+                            </a>
+                        </td>
                         <td>Name</td>
                         {columnNames.map(name => (
                             <td key={name}>
-                                {name}
+                                <a onClick={(event) => {this.handleSortTable(event, name)}}>
+                                    {name}
+                                </a>
                                 <input
                                     onClick={(event) => {this.handleShowHeatmap(event, name)}}
                                     type="checkbox"
