+++ src/js/tableData/Layout.js
@@ -5,7 +5,6 @@ import DataTable from './DataTable';
 import PlateGrid from '../plate/PlateGrid';
 import Dataset from '../dataset/Dataset';
 import Footer from '../Footer';
-import clusterfck from 'clusterfck';
 
 export default React.createClass({
 
@@ -66,74 +65,11 @@ export default React.createClass({
         }
     },
 
-    clusterTableData: function() {
-        // make a list of data for each image from state.tableData
-        // Each item of tableData is already a dict of {imageId: value}
-        let dataKeys = Object.keys(this.state.tableData);
-        if (dataKeys.length < 2) return;
-        // make dict of {imageId: [v1, v2, v3]}.
-        let imageIds = this.props.filteredImages.map(i => i.id);
-        let toCluster = imageIds.map(iid => {
-            return dataKeys.map(key => this.state.tableData[key][iid]);
-        });
-        // Make a lookup of tableDataKey :image_id so that when we get clustered tableDAta
-        // we can work out which image each 'row' of data came from (since we can't
-        // include image ID in the cluster data)
-        // We know the 'toCluster' list is in same order as imageIDs just now...
-        let clusterLookup = imageIds.reduce((prev, iid, index) => {
-            let dataString = toCluster[index].join(",");
-            prev[dataString] = iid;
-            return prev;
-        }, {});
-        // let toCluster = [[255, 255, 240],
-        //     [20, 120, 102],
-        //     [250, 255, 253],
-        //     [100, 54, 300]];
-        let threshold = 25000;
-        var clusters = clusterfck.hcluster(toCluster, clusterfck.EUCLIDEAN_DISTANCE,
-            clusterfck.AVERAGE_LINKAGE, threshold);
-        console.log(clusters);
-        let orderedResults = this.traverseCluster(clusters);
-
-        let orderedImageIds = orderedResults.map(res => clusterLookup[res.join(",")]);
-        console.log('orderedImageIds', orderedImageIds);
-        return orderedImageIds;
-    },
-
-    traverseCluster: function(clusters) {
-
-        let out = [];
-        function traverseNode(node) {
-            console.log("traverse", node);
-            if (node.left && node.right) {
-                traverseNode(node.left);
-                traverseNode(node.right);
-            } else {
-                out.push(node.value);
-            }
-        }
-        clusters.forEach(n => {traverseNode(n)});
-        console.log(out);
-        return out;
-    },
-
     render: function() {
         if (this.props.plateData === undefined && this.props.filteredImages === undefined) {
             return(<div></div>)
         }
-        let orderedImageIds = this.clusterTableData();
         let filteredImages = this.props.filteredImages;
-        if (orderedImageIds) {
-            // sort filteredImages by ID...
-            // create lookup {iid: image} 
-            console.log("BEFORE CLUSTER", filteredImages.map(img => img.id));
-            let imgLookup = filteredImages.reduce((prev, img) => {
-                prev[img.id] = img;
-                return prev;
-            }, {});
-            filteredImages = orderedImageIds.map(iid => imgLookup[iid]);
-            console.log("AFTER CLUSTER", filteredImages.map(img => img.id));
-        }
         let imageComponent;
         if (this.state.layout === "table") {
             imageComponent = (
